LayoutTests/imported/w3c/web-platform-tests/reporting/cross-origin-report-no-credentials.https.sub-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'report-to'.

CONSOLE MESSAGE: Refused to load https://localhost:9443/reporting/resources/fail.png because it does not appear in the img-src directive of the Content Security Policy.


FAIL Reporting endpoints did not receive credentials. assert_unreached: A report of csp-violation from https://localhost:9443/reporting/cross-origin-report-no-credentials.https.sub.html is not found. Reached unreachable code
